+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/scene/interactive_scene.py
@@ -16,7 +16,7 @@ from pxr import PhysxSchema
 
 import omni.isaac.orbit.sim as sim_utils
 from omni.isaac.orbit.assets import Articulation, ArticulationCfg, AssetBaseCfg, RigidObject, RigidObjectCfg
-from omni.isaac.orbit.sensors import FrameTransformerCfg, SensorBase, SensorBaseCfg
+from omni.isaac.orbit.sensors import ContactSensorCfg, FrameTransformerCfg, SensorBase, SensorBaseCfg
 from omni.isaac.orbit.terrains import TerrainImporter, TerrainImporterCfg
 
 from .interactive_scene_cfg import InteractiveSceneCfg
@@ -348,6 +348,11 @@ class InteractiveScene:
                         target_frame.prim_path = target_frame.prim_path.format(ENV_REGEX_NS=self.env_regex_ns)
                         updated_target_frames.append(target_frame)
                     asset_cfg.target_frames = updated_target_frames
+                elif isinstance(asset_cfg, ContactSensorCfg):
+                    updated_filter_prim_paths_expr = []
+                    for filter_prim_path in asset_cfg.filter_prim_paths_expr:
+                        updated_filter_prim_paths_expr.append(filter_prim_path.format(ENV_REGEX_NS=self.env_regex_ns))
+                    asset_cfg.filter_prim_paths_expr = updated_filter_prim_paths_expr
 
                 self._sensors[asset_name] = asset_cfg.class_type(asset_cfg)
             elif isinstance(asset_cfg, AssetBaseCfg):